requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
python-jose[cryptography]==3.3.0
bcrypt==5.0.0
boto3==1.34.0
openai==1.30.0
python-dotenv==1.0.1
pydantic-settings==2.2.1
httpx==0.27.0
Pillow==10.3.0
pdf2image==1.17.0
